refactor(stores): rename Zustand store files to useXStore naming convention
- auth.store.ts    → useAuthStore.ts
- board.store.ts   → useBoardStore.ts
- praises.store.ts → usePraisesStore.ts
- socket.store.ts  → useSocketStore.ts
- ui.store.ts      → useUiStore.ts
